services/ai-engine/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
influxdb==5.3.1
paho-mqtt==1.6.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
python-multipart==0.0.6
streamlit==1.29.0
supervisor==4.2.5
